Timesheet EDA with R
This script performs a simple transform of data in a wide format to a long format, plus some cleanup for initial Exploratory Data Analysis (EDA). The data is in a spreadsheet generated with LibreOffice Calc (ODS format). It can run in RStudio on Windows or Ubuntu - make sure to specify the correct path to the ODS file.