Sample Monte Carlo simulation with ggplot
This simulation was inspired by the presentation in the last Twin Cities R user group. I made this simulation for a one-person scenario to see what are the chances of getting to a place on time or early.
It also showcases the use of ggplot2 package to create a better histogram visualization.