Flask app to estimate trip using Monte Carlo
This was inspired by a presentation in the Twin Cities R Meetup where the presenter showcased Monte Carlo to calculate the chances of on-time arrival. 
I made a similar calculation using Python and Flask to allow users to key-in their data, and Flask returns the result along with the histogram.
This is a very simple Flask app - you can expand it in a more robust fashion using HTML and CSS with proper data entry checking and logging.